fix zombie movement, block placement

+++ rd-160052/include/character/ZombieModel.hpp
@@ -14,5 +14,5 @@ class ZombieModel {
 
         ZombieModel();
 
-        void render(float time);
+        void render(double time);
 };
